Create object_detection.py
This is merely a draft of the program for safe keeping, has not been cleaned up and a lot of dependencies have not been added to the Gitthub repo (yet). Example is the grab_screen utility.

This is also currently not modular. Will not work on your system until modularity is added.